Many many changes to the layout.  Highlights:
Added diodes to prevent the batteries from charging each other
Added an off switch to isolate the batteries.
Fixed the size of the battery.  It's larger than it was supposed to be.
Bottom battery module ports are now available.
Switched to a surface mount crystal
Switched to contact-only programming headers.  Pitch is the same, but it's only SMD now.
Opened up some space for the larger microprocessor.